Module net_neurons::nnetwork

Modules

loss_functions

Structs

CalcNode

Wrapper class for CalcNodeCore. This is the struct intended to be used.

CalcNodeCore

The struct that actually holds the data. Only a few member functions are implemented for this struct, use the wrapper struct CalcNode instead.

FunctionLayer

A Layer type that applies a mathematical function to the input CalcNode. The function is responsible for also setting up the gradient calculation for back propagation to work.

LinearLayer

Standard layer consisting of a matrix of weights and a column vector of biases. The biases are optional.

MultiLayer

This struct is just a stack of Layers with some convenience attached.

ReshapeLayer

A Layer type that only coerces the input into a new shape without changing its values.

Enums

NodeType

Tensors of higher dimensions than matrices are not supported

VecOrientation

Mostly used for debug output

Traits

Layer

Parameters

Object implementing this trait must supply iterators to all its parameters, in arbitrary, but fixed, order.

Type Aliases

FloatType

If we ever want to change to f32, it is easy

NodeShape

All nodes are treated as matrices of shape (rows, columns)