Struct net_neurons::nnetwork::CalcNode

pub struct CalcNode { /* private fields */ }

Wrapper class for CalcNodeCore. This is the struct intended to be used.

Implementations

impl CalcNode

pub fn len(&self) -> usize

Returns the size, i.e., the number of values in the node

pub fn is_empty(&self) -> bool

impl CalcNode

pub fn node_type(&self) -> NodeType

Returns the enum NodeType to easily categorize what kind of node we are dealing with.

impl CalcNode

pub fn copy_parents(&self) -> Vec<CalcNode>

The returned CalcNode copies are shallow. The underlying CalcNodeCores are not copied.

pub fn shape(&self) -> NodeShape

pub fn copy_vals(&self) -> Vec<FloatType>

Returns a Vec containing the raw float values.

pub fn copy_grad(&self) -> Vec<FloatType>

Returns a Vec containing the raw float gradient values.

pub fn value_indexed(&self, i: usize) -> FloatType

Returns a specific value by index

pub fn gradient_indexed(&self, i: usize) -> FloatType

Returns a specific gradient value by index

pub fn set_vals(&mut self, vals: &[FloatType])

Overwrites all values with the ones supplied. Will panic if a wrong number of values are given.

pub fn set_value_indexed(&mut self, i: usize, val: FloatType)

Sets one specific value by index.

pub fn add_grad(&mut self, grad: &[FloatType])

Increments all gradients element-wise with the values supplied.

pub fn reset_grad(&mut self)

Resets all gradient values to zero.

pub fn set_grad(&mut self, grad: &[FloatType])

Overwrites all gradient values with the ones supplied. Will panic if a wrong number of values are given.

pub fn set_gradient_indexed(&mut self, i: usize, val: FloatType)

Sets one specific gradient by index.

impl CalcNode

pub fn new( shape: NodeShape, vals: Vec<FloatType>, parents: Vec<CalcNode>, back_propagation: Option<Box<dyn Fn(CalcNode)>>, ) -> Self

Returns a new instance of a CalcNode containing a CalcNodeCore with the supplied member data. Will panic if the size of the NodeShape and data are not equal.

Example 1

use net_neurons::nnetwork::CalcNode;
 
let shape = (1,1);
let vals = vec![1.];
let parents = vec![];
let node1 = CalcNode::new(shape, vals, parents, None);

let shape = (1,1);
let vals = vec![1.];
let parents = vec![node1];
let node2 = CalcNode::new(
    shape,
    vals,
    parents,
    Some( Box::new( |child_node: CalcNode| {
        for mut p in child_node.copy_parents(){
            p.add_grad(&vec![1.]);
        }
    })));

pub fn new_scalar(value: FloatType) -> Self

Shortcut for constructing scalar nodes.

pub fn new_col_vector(value: Vec<FloatType>) -> Self

Shortcut for constructing column vector nodes.

pub fn new_row_vector(value: Vec<FloatType>) -> Self

Shortcut for constructing row vector nodes.

pub fn new_from_shape(shape: NodeShape, vals: Vec<FloatType>) -> Self

Constructs a node with a specific shape and values. Will panic if the size of the shape and data are not equal.

Example

use net_neurons::nnetwork::CalcNode;
 
let shape = (2,2); // a 2 by 2 matrix
let vals = vec![1.,2.,3.,4.];
let node = CalcNode::new_from_shape(shape, vals);

pub fn rand_from_shape(shape: NodeShape) -> Self

Similar to CalcNode::new_from_shape but randomizes all values from a normal distribution.

pub fn reshape(&mut self, shape: NodeShape)

Coerce the node into a new shape. Will panic if the size changes.

impl CalcNode

pub fn sum(&self) -> CalcNode

Returns the sum of all values in the CalcNode, i.e., the result will be a scalar.

pub fn normalized(&self) -> CalcNode

Normalises the CalcNode so that its sum becomes unity.

pub fn collapse(&self) -> CalcNode

Collapses the CalcNode to a one-hot vector. Only makes sense if you can interpret the values as a probabilities that sum to unity.

impl CalcNode

pub fn inv(&self) -> CalcNode

Inverts all values.

impl CalcNode

pub fn exp(&self) -> CalcNode

Exponentiates all values.

impl CalcNode

pub fn log(&self) -> CalcNode

Applies the natural logarithm to all values.

impl CalcNode

pub fn pow(&self, power: &CalcNode) -> CalcNode

Applies the power function to all values.

impl CalcNode

pub fn element_wise_mul(&self, other: &Self) -> CalcNode

Multiplies the two objects element-wise.

Multiplication by scalar and the dot product are provided by CalcNode::mul.

impl CalcNode

pub fn back_propagation(&mut self)

Recalculates the gradients of all nodes leading up to the current one. The gradient of the current node is set to unity.

Example

use net_neurons::nnetwork::CalcNode;
 
let a = CalcNode::new_scalar(2.);
let mut b = &a * &a;
b.back_propagation();
assert_eq!(a.gradient_indexed(0), 4.);
assert_eq!(b.gradient_indexed(0), 1.);

pub fn decend_grad(&mut self, learning_rate: FloatType)

Decends the gradient by a fraction of the calculated gradient.