Struct net_neurons::nnetwork::MultiLayer

pub struct MultiLayer { /* private fields */ }

This struct is just a stack of Layers with some convenience attached.

Implementations

impl MultiLayer

pub fn new(layers: Vec<Box<dyn Layer>>) -> Self

All Layers are constructed beforehand and then put into the MultiLayer.

pub fn set_loss_function(&mut self, f: &'static LossFuncType)

Use your own or one of the ones provided in the module crate::nnetwork::mlp::loss_functions.

pub fn set_regularization(&mut self, reg: Option<FloatType>)

Set to Some(float) to punish non-zero parameters.

pub fn len(&self) -> usize

Returns the number of Layers

pub fn get_layer(&self, i: usize) -> &dyn Layer

Returns the indexed layer. Panics if the index invalid.

pub fn forward(&self, inp: &CalcNode) -> CalcNode

Forwards the input through all Layers and returns the final result.

pub fn loss(&self, inp: &[(CalcNode, CalcNode)]) -> CalcNode

Calculates the average loss of the list of (prediction, truth) tuples.

pub fn train( &mut self, inp: &[(CalcNode, CalcNode)], learning_rate: FloatType, ) -> FloatType

Trains the network on the supplied training data and returns the average loss.

pub fn export_parameters(&self, filename: &str) -> Result<String>

Exports all parameters to a text file.

The filename is returned upon successful export.

Adds a numerical suffix if the wanted filename is taken.

pub fn import_parameters(&mut self, filename: &str) -> Result<(), Error>

Imports previously exported parameters. Expect errors if the current network does not have the same setup of layers.