Trait net_neurons::nnetwork::Layer

pub trait Layer: Parameters + Display {
    // Required methods
    fn forward(&self, inp: &CalcNode) -> CalcNode;
    fn layer_name(&self) -> &str;

    // Provided method
    fn shape(&self) -> Option<NodeShape> { ... }
}

Required Methods

fn forward(&self, inp: &CalcNode) -> CalcNode

Calculates the output given an input.

fn layer_name(&self) -> &str

All Layers must have a name

Provided Methods

fn shape(&self) -> Option<NodeShape>

The shape determines what input shapes the layer accepts as well as the output shape it will give. For some Layers it makes no sense to have a shape. If it makes sense though, this function must be overidden.