Struct net_neurons::retext::CharSet

pub struct CharSet { /* private fields */ }

Holds the data consisting of lines of text. Also holds an ordered set of all characters by which the CharSet encodes and decodes characters to and from CalcNodes.

Implementations

impl CharSet

pub fn new(path: &str, training_ratio: f32, lowercase: bool) -> Self

Loads the data from the text file and creates the character set. Optionally all characters are made lowercase. The training ratio decides how much of the data that is available for training and not reserved for validation.

Example

use net_neurons::retext::CharSet;
 
let ds1 = CharSet::new("./datasets/tiny_shakespeare.txt", 1., true);
assert_eq!(ds1.number_of_chars(), 26);
let ds2 = CharSet::new("./datasets/tiny_shakespeare.txt", 1., false);
assert_eq!(ds2.number_of_chars(), 52);

pub fn add_character(&mut self, c: char)

Add extra characters to the set, e.g., sentinels.

pub fn characters(&self) -> &[char]

Returns a slice of all currently known characters in the set.

Example

use net_neurons::retext::CharSet;
 
let ds = CharSet::new("./datasets/test.txt", 1., true);
assert_eq!(ds.characters(), &['a','b','c']);

pub fn training_data(&self) -> &[String]

Returns a slice of all available training data.

pub fn validation_data(&self) -> &[String]

Returns a slice of all available validation data.

pub fn number_of_chars(&self) -> usize

Number of known characters.

pub fn decode_char(&self, vector: &CalcNode) -> Result<char, DataSetError>

Interprets a CalcNode as a character. The CalcNode is assumed to be a one-hot vector.

Example

use net_neurons::{nnetwork::CalcNode,retext::CharSet};
 
let ds = CharSet::new("./datasets/test.txt", 1., true);
let cn = CalcNode::new_col_vector(vec![0., 0., 1.]);
assert_eq!( ds.decode_char(&cn), Ok('c') );

pub fn decode_string(&self, v: &[&CalcNode]) -> Result<String, DataSetError>

Calls CharSet::decode_char for every CalcNode] in the slice, and returns a string. The CalcNodes are assumed to be a one-hot vectors.

pub fn encode(&self, s: &str) -> Result<CalcNode, DataSetError>

Creates a matrix where each column is a one-hot vector which corresponds a character in the string.

Example

use net_neurons::{nnetwork::CalcNode,retext::CharSet};
 
let ds = CharSet::new("./datasets/test.txt", 1., true);
let cn = ds.encode("b").unwrap();
assert_eq!(cn.value_indexed(0),0.);
assert_eq!(cn.value_indexed(1),1.);
assert_eq!(cn.value_indexed(2),0.);