Struct net_neurons::retext::ReText

pub struct ReText { /* private fields */ }

Manages a network that predicts the next character in a name based on the x previous characters.

Implementations

impl ReText

pub fn new( data: CharSet, block_size: usize, embed_dim: Option<usize>, n_hidden_layers: usize, layer_dim: usize, regularization: Option<FloatType>, ) -> ReText

The block_size determines how many characters are used to predict the next one.

After each linear layer a non-linear FunctionLayer is inserted. The last layer is a softmax calculation.

pub fn train( &mut self, cycles: usize, learning_rate: FloatType, batch_size: usize, verbose: bool, )

Trains the network for the specified number of cycles. Each cycles uses ´batch_size´ data points. The learning rate is a constant for all cycles.

pub fn predict( &mut self, seed_string: &str, n_char: usize, ) -> Result<String, DataSetError>

Predicts the next n_char characters given the starting `seed_string´. If a sentinel token is predicted the string is terminated regardless of size.

pub fn characters(&self) -> &[char]

pub fn export_parameters(&self, filename: &str) -> Result<String>

pub fn import_parameters(&mut self, filename: &str) -> Result<(), Error>